fix: bug-0031 — scale simple-font /Widths and FontDescriptor metrics to glyph space (v0.11.11)
The simple-font path emitted /Widths and every FontDescriptor metric in the
embedded font's raw unitsPerEm, but PDF glyph space is 1000 units/em — so any
embedded font whose upem was not 1000 (Arial, Verdana, most macOS TrueType
faces use 2048) laid text out upem/1000x too wide.  font_helpers now scales
every advance and descriptor metric by 1000/upem via new glyph_space_scale +
scale_metric helpers, the same formula the composite /W path already used;
FontDescriptorPdfInfo's metric fields widened i16 -> i32 so a small-upem font
can't overflow when scaled up.  New regression test
tests/font_metrics_scaling_regression.rs, verified to fail on revert.

+++ medpdf/src/pdf_watermark.rs
@@ -288,7 +288,7 @@ fn widths_as_object_array(widths: &[u16]) -> Object {
     Object::Array(widths.iter().map(|v| Object::Integer(*v as i64)).collect())
 }
 
-fn bbox_as_object_array(bbox: &[i16]) -> Result<Object> {
+fn bbox_as_object_array(bbox: &[i32]) -> Result<Object> {
     if bbox.len() != 4 {
         return Err(MedpdfError::new("FontBBox must have exactly 4 elements"));
     }
